requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albucore=0.0.33=pypi_0
albumentationsx=2.0.9=pypi_0
annotated-types=0.7.0=pypi_0
aom=3.6.0=h6a678d5_0
asttokens=3.0.0=pyhd8ed1ab_1
blas=1.0=mkl
bottleneck=1.4.2=py310ha9d4c09_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
cairo=1.18.4=h44eff21_0
certifi=2025.6.15=pypi_0
charset-normalizer=3.4.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
dav1d=1.2.1=h5eee18b_0
debugpy=1.8.14=py310hf71b8c6_0
decorator=5.2.1=pyhd8ed1ab_0
einops=0.8.1=pypi_0
einops-exts=0.0.4=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.1=h6a678d5_0
fairscale=0.4.13=pypi_0
filelock=3.18.0=pypi_0
fontconfig=2.15.0=h2c49b7f_0
fonttools=4.58.5=pypi_0
freetype=2.13.3=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
fsspec=2025.5.1=pypi_0
geopandas=1.1.1=pypi_0
gigapath=0.1.0=pypi_0
graphite2=1.3.14=h295c915_1
h5py=3.14.0=pypi_0
harfbuzz=10.2.0=hdfddeaa_1
hf-xet=1.1.5=pypi_0
huggingface-hub=0.33.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.37.0=pyh8f84b5b_0
ipywidgets=8.1.7=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
joblib=1.4.2=py310h06a4308_0
jpeg=9f=h5ce9db8_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
jupyterlab-widgets=3.0.15=pypi_0
kiwisolver=1.4.8=py310h6a678d5_0
krb5=1.21.3=h143b758_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libavif=1.1.1=h5eee18b_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.84.2=h37c7471_0
libgomp=15.1.0=h767d61c_3
libiconv=1.16=h5eee18b_3
libpng=1.6.39=h5eee18b_0
libsodium=1.0.20=h4ab18f5_0
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxml2=2.13.8=hfdd30dd_0
llvmlite=0.44.0=py310hc1e8f15_1
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-base=3.10.6=py310h71dd3da_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
numba=0.61.2=py310h8648a56_1
numexpr=2.11.0=py310hf4471f9_0
numpy=2.2.6=pypi_0
numpy-base=2.0.1=py310hb5e798b_1
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.12.0.88=pypi_0
openjpeg=2.5.2=h0d4d230_1
openslide-bin=4.0.0.8=pypi_0
openslide-python=1.4.2=pypi_0
openssl=3.5.2=h26f9b46_0
packaging=25.0=pyh29332c3_1
pandas=2.3.0=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=py310hb1c3d2d_0
pip=25.1=pyhc872135_2
pixman=0.46.4=h7934f7d_0
platformdirs=4.3.8=pyhe01879c_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py310ha75aee5_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pynndescent=0.5.13=pyhd8ed1ab_1
pyogrio=0.11.0=pypi_0
pyparsing=3.2.3=pypi_0
pyproj=3.7.1=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2025.2=py310h06a4308_0
pyyaml=6.0.2=pypi_0
pyzmq=27.0.0=py310h71f11fc_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.4=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.7.2=py310h505adc9_0
scipy=1.15.3=py310h525edd1_0
seaborn=0.13.2=py310h06a4308_3
segmentation-models-pytorch=0.5.0=pypi_0
setuptools=78.1.1=py310h06a4308_0
shapely=2.1.1=pypi_0
simsimd=6.5.0=pypi_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
stringzilla=3.12.6=pypi_0
sympy=1.14.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py310h2f386ee_0
timm=0.9.16=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.21.2=pypi_0
torch=2.7.1=pypi_0
torchvision=0.22.1=pypi_0
tornado=6.5.1=py310ha75aee5_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.53.1=pypi_0
trident=0.2.0=pypi_0
triton=3.3.1=pypi_0
typing-extensions=4.14.0=pypi_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
umap-learn=0.5.9.post2=py310hff52083_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py310h06a4308_0
widgetsnbextension=4.0.14=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxrender=0.9.12=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h3b0a872_7
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
